example/python/modbus_slave_full/pyproject.toml
[tool.poetry]
name = "modbus-slave-full"
version = "1.0.0"
description = "功能完整的 Modbus TCP/RTU 服务器，支持 Web 控制台"
authors = ["clint <clinton_luO@163.com>"]
readme = "README.md"
license = "MIT"
keywords = ["modbus", "tcp", "rtu", "server", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: System :: Hardware",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
aiohttp = "^3.9.0"
pyserial-asyncio = "^0.6"
pyyaml = "^6.0"
aiohttp-cors = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[tool.poetry.scripts]
modbus-server = "modbus_slave_full.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
